refactor: Chat 메시지 관련 로직 커스텀훅으로 분리
